refactor based on review feedback
